refactor: move config resolvers to shared package
