+++ src/quickapp/internal_tooling/internal_tooling_module.py
@@ -9,7 +9,7 @@ from quickapp.common.tool_names import INTERNAL_CODE_EXECUTION_PYTHON_INTERPRETE
 from quickapp.config.application import ApplicationConfig
 from quickapp.config.tools.predefined import PredefinedTool
 from quickapp.config.toolsets.internal import InternalToolSet
-from quickapp.config_resolvers.tool_timeout_resolver import ToolTimeoutResolver
+from quickapp.shared.config_resolvers.tool_timeout_resolver import ToolTimeoutResolver
 from quickapp.internal_tooling.py_interpreter_tooling._py_interpreter_client import (
     _PyInterpreterClient,
 )
