+++ src/quickapp/shared/config_resolvers/stage_display_settings.py
