+++ src/tests/unit_tests/config_resolvers_tests/test_file_loading_size_limit_resolver.py
@@ -4,8 +4,8 @@ import pytest
 from pydantic import ValidationError
 
 from quickapp.config.application import Features, FileLoadingConfig
-from quickapp.config_resolvers.file_loading_settings import FileLoadingSettings
-from quickapp.config_resolvers.file_loading_size_limit_resolver import FileLoadingSizeLimitResolver
+from quickapp.shared.config_resolvers.file_loading_settings import FileLoadingSettings
+from quickapp.shared.config_resolvers.file_loading_size_limit_resolver import FileLoadingSizeLimitResolver
 
 
 def _resolver(env: int, app: int | None) -> FileLoadingSizeLimitResolver:
