+++ src/tests/unit_tests/dial_core_services_tests/test_dial_file_service.py
@@ -4,7 +4,7 @@ import pytest
 from aidial_client._exception import EtagMismatchError
 
 from quickapp.common.state_holder import StateHolder
-from quickapp.config_resolvers.file_loading_size_limit_resolver import FileLoadingSizeLimitResolver
+from quickapp.shared.config_resolvers.file_loading_size_limit_resolver import FileLoadingSizeLimitResolver
 from quickapp.dial_core_services.dial_file_service import DialFileService
 
 DEFAULT_LIMIT = 10 * 1024 * 1024
